feat: add stock quantity in product, add validation based on stock quantity in controllers and models

+++ app/controllers/api/v1/carts_controller.rb
@@ -2,13 +2,11 @@
 
 module Api
   module V1
-    # Контроллер для управления корзинами пользователя
     class CartsController < BaseController
       before_action :authenticate_user!
 
       # Получить все активные корзины текущего пользователя
       def index
-        # todo: выяснить в чем ошибка с :cart_items
         carts = policy_scope(Cart).active.includes(:cart_items)
         render json: carts, each_serializer: CartSerializer, status: :ok
       end
@@ -39,7 +37,8 @@ module Api
 
         authorize cart
 
-        product = Product.find(params[:product_id])
+        data = cart_params
+        product = Product.find(data[:product_id])
 
         if product.shop_id != cart.shop_id
           return render_error(
@@ -50,17 +49,35 @@ module Api
           )
         end
 
-        # Сначала обязательно сохраняем корзину, чтобы получить cart.id
+        quantity_to_add = data[:quantity].to_i
+
+        if quantity_to_add <= 0 || quantity_to_add > 100
+          return render_error(
+            key: "cart.invalid_quantity",
+            message: "Количество товара должно быть от 1 до 100",
+            code: :unprocessable_entity,
+            status: :unprocessable_entity
+          )
+        end
+
         unless cart.persisted?
           return render_validation_errors(cart) unless cart.save
         end
 
-        # Теперь можно безопасно работать с cart_items
         cart_item = cart.cart_items.find_or_initialize_by(product: product)
         cart_item.quantity ||= 0
-        cart_item.quantity += params[:quantity].to_i.clamp(1, 100)
+
+        if cart_item.quantity + quantity_to_add > product.stock_quantity
+          return render_error(
+            key: "cart.insufficient_stock",
+            message: "Недостаточно товара на складе. Доступно: #{[ product.stock_quantity - cart_item.quantity, 0 ].max}",
+            code: :unprocessable_entity,
+            status: :unprocessable_entity
+          )
+        end
+
+        cart_item.quantity += quantity_to_add
         cart_item.price_snapshot = product.price
-        cart_item.cart_id ||= cart.id # на всякий случай
 
         if cart_item.save
           render json: cart, serializer: CartSerializer, status: :created
@@ -68,8 +85,6 @@ module Api
           render_validation_errors(cart_item)
         end
       end
-
-
 
       # Удалить товар из корзины
       def remove_item
@@ -84,7 +99,8 @@ module Api
 
         authorize cart
 
-        product = Product.find(params[:product_id])
+        data = cart_params
+        product = Product.find(data[:product_id])
         cart_item = cart.cart_items.find_by(product_id: product.id)
         return render_error(
           key: "cart_item.cart_item_not_exists_in_cart",
@@ -93,7 +109,16 @@ module Api
           status: :not_found
         ) unless cart_item
 
-        quantity_to_remove = params[:quantity].to_i.clamp(1, 100)
+        quantity_to_remove = data[:quantity].to_i
+
+        if quantity_to_remove <= 0 || quantity_to_remove > 100
+          return render_error(
+            key: "cart.invalid_quantity",
+            message: "Количество для удаления должно быть от 1 до 100",
+            code: :unprocessable_entity,
+            status: :unprocessable_entity
+          )
+        end
 
         if cart_item.quantity <= quantity_to_remove
           cart_item.destroy
@@ -107,13 +132,16 @@ module Api
 
       private
 
-      # Возвращает корзину по shop_id или создает новую
+      def cart_params
+        params.require(:cart).permit(:product_id, :quantity)
+      end
+
       def find_or_initialize_cart
         user = current_user
         return render_unauthorized unless user
 
         Cart.find_or_initialize_by(user_id: user.id, shop_id: params[:shop_id]).tap do |cart|
-          cart.expired_at ||=30.days.from_now
+          cart.expired_at ||= 30.days.from_now
         end
       end
     end
